+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 ## Fixed Issues
 
 - [core] Re-enable the circuit breakers for the destination and xsuaa services.
+- [odata] Fix encoding of filter strings.
 
 
 # 1.44.0